requirements.txt
matplotlib==3.7.1
numpy==1.23.5
Pillow==10.0.0
torch==2.0.1
torchvision==0.15.2
